backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
pymongo==4.6.1
djongo==1.3.6
